clone.sh: only checkout topic for other repos
If a Gerrit topic is available, the clone script will always checkout
the tip of the topic in all the repositories. This means that when a
Gerrit review triggers the CI, the full topic will always be verified,
which may cause the review that triggered the CI to fail the
verification because there was an issue with one of the later reviews in
the topic chain.

To make it possible to verify a specific review in the topic chain, the
clone script will now checkout the specific review that triggered the
job and then only use the topic to checkout the other repositories.

Signed-off-by: Mikael Olsson <mikael.olsson@arm.com>
Change-Id: Icf747351b6cd9f0980840179148db22eed6638e2