tf-a-daily: Switch to run at 'H 2 * * *'
Effectively, switch start time to that of tf-a-daily, but space them both
by +1hr each. So, tf-m-nightly starts at 16:HH, tf-a-daily - 22:HH. This
is based on the idea that TF-M is being worked largely by Chinese and
European teams, while TF-A - by European and US teams, so try to use
accommodating times.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I6c0dba6d2da6d864110c89c09bf677033385f995