tf-a-gerrit-tforg-l2: run tf-plat-tests tests group
The tf-plat-tests group is meant to be a set of tests configurations
for running self-tests on the TF-A boot images (BL1 and BL2). Right
now, we have 2 of them:

 1) the delegated attestation and measurement tests introduced in TF-A
    commit 25dd2172ae564c7 "feat(tc): add delegated attest and
    measurement tests".

 2) the RSS/AP non-volatile counters introduced in TF-A commit
    1b0761130bab903 "test(tc): test for AP/RSS NV counter interface".

Only 1) is enabled in the OpenCI.
2) is pending integration into the OpenCI.
More tests will be added in the future.

Both target the Arm TC2 FVP platform.

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>
Change-Id: I8ccac9ee20cff4d7d9005aa067203df629d7db2e