tf-a-ci-gateway: Increase timeout to 4h, allowing to wait for more subjobs
tf-a-ci-gateway doesn't do much beyond waiting for tf-a-builder subjobs.
And with a loaded system, individual tf-a-builder jobs may take longer to
finish. E.g. https://ci.trustedfirmware.org/job/tf-a-ci-gateway/37359/console
where individual tf-a-builder's were sucessful, but took long time, and
tf-a-ci-gateway failed due to timeout without catching the rest of them.
So, increase the timeout by gradual amount.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I2b736609125725974aaa10a829afcebdf83383db