tf-a-builder: Don't archive artefacts-lava/, just record BIN_MODE instead
artefects-lava is a symlink created as part of internal CI processing,
to simplify means of selecting which configuration (release/debug) to
post to LAVA. The only reason to archiving that symlink as a job
artifacts is also recording this selection for reviewer of the build
results. But Jenkins actually doesn't archive symlinks as is, but copies
the underlying directories/files. So, the connection of artefacts-lava
to release/debug binaries is actually lost, it's just another (third!)
set of binaries, and would require scrutiny to understand which binaries
it's copy of.

Summing up, there's no good to reason to archive artefacts-lava at all.
So, we skip doing that, and instead record binary configuration submitted
to LAVA to the lava-binmode.txt file in case a human reviewer needs that
information. As the end result, we make it more clear, and save 1/3 of
binary space (problems with the disk space what originally prompted to
look into these issues).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: If70c70ee57e7c5a75f8eec65156d11c9b57141c3