clone.sh: When calculating REPO_REFSPEC for topic, don't swallow errors
There can be unlimited number of unexpected errors, e.g. we had a case when
SSH failing due to key type mismatch (TFC-459). Suh should not be swallowed,
as that can be potentially lead to a case when completely different refspec
would be tested than expected. Instead, any expected conditions when
REPO_REFSPEC may lead to not a valid value (to be replaced with a default)
should be handled explicitly.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I35078258eef051306f45d18369f32b74e7bc36a2