tf-a-builder: Add CLEAN_UP_ARTIFACTS param, default to 1
If the value is "0", all binary artifacts are archived, as before. If it's
"1", remove *.elf files, as they're not really needed for LAVA testing. The
idea of setting it to 0 by default is that 99% of test results just go to
LAVA, so we don't need to store ELFs. In case a particular build run needs
to be investigated and ELFs are useful, it can be rebuilt with
CLEAN_UP_ARTIFACTS=0 to get them.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I085ef69e49314c73186f8c785c557b82998b8933