tf-a-commitlint: Upgrade to Jammy Docker image
This includes workaround for issue with running "nvm", etc. Node tools.
They didn't work as expected, because environment for them wasn't properly
set up, which is in turned happened because somehow Jammy's bash thought
that Jenkins pipeline runs it in interactive mode, so it didn't evaluate
BASH_ENV on startup, and instead evaluated .bashrc files.

All this is of course confusing and esoteric. And the root cause lies in
using Jenkins pipeline feature for this otherwise trivial job. The job
doesn't require any pipeline features, and instead a lot of effort
(and obscure workarounds) are spent on a trivial task of maintaining
consistent environment across multiple shell commands - something
which would be trivial with a shell script and "freestyle" (i.e.,
normal) Jenkins job.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I6a292c76419fad062ff50921ca1500d0a471d153