TF-A.yaml: Filter queue and executors to match the TF-A view
Our Jenkins server is shared with large number of jobs of different
projects, oftentimes running in parallel. It gets unwieldy to watch
all them shown together. So, the option to filter build queue and
executors to a partical subset of jobs, when using a particular
job view - makes really good sense.

This is similar to change applied recently to the TF-M view.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I0d3b85440851e4e2383692e3ae1f3b4618155a77