tf-a-eclair-daily: Remove old version of the job.
This job was already backed up as tf-a-eclair-daily-separate. Otherwise,
it's going to be replaced with a new version of the job producing cumulative
report across all test configurations. This involves changing the Jenkins
job type (from matrix to freestyle), and such cases are known to require
explicit delete/recreate with Jenkins Job Builder.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Ia7217bb1759aeff8f2bca1c9c118ca969b504aff