tf-a-gerrit-tforg-l2: Set LAVA_RETRIES to 2
This job is supposed to be "interactive", providing more or less quick
feedback to patch submitters/reviewers. In that regard, it makes no
sense to repeat legitimately failing tests multiple times - they still
will be failing. And worst case, if failures is by timeout, such
repetitions will take too much time, causing bottleneck in the LAVA
queue, and knock-on effect on other builds (which may take too long
and/or timeout).

On the other hand, we're still (or at least used to be) limited by some
non-determistically failing tests which need multiple times to pass, so
still have 2 retries by default.

Going forward, we need to enumerate and isolate non-determistically failing
tests, retrying only them, and not all tests blindly.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Iaf10176c5a2facae8fcc2c30c389897a68a7d42d