tf-a-static-checks: Increase timeout by 1hr
We had a case https://ci.trustedfirmware.org/job/tf-a-static-checks/6760/
when tf-a-static-checks timed out after 3hrs, while underlying tf-a-ci-gateway
subjob finished successfully in 2h34m. This is apparently due to high load
of the Jenkins server, where subjobs spend too much time in queue (also,
there's suspicion that due to high load, Jenkins "doesn't notice" when a
subjob completes in a due time). Either way, this seems to be the case when
increasing timeouts may actually help, not harm.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I3f2ca8e1bdd0e125ebf3c0438e2275184764c118