tf-a-daily: move windows build to main job
The windows build job only tests two build configurations. It makes more
sense to have it nested in the main job with other build configurations
and not at the top level. Having it there implies there is some special
consideration for windows builds when we in-fact treat it like any other
build job.

Change-Id: Ia9024694cf8b21749c4125d04b79547a28a0dcbb
Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>