tf-a-daily.yaml: rearrange jobs
tf-a-daily triggers tf-a-main, tf-a-windows-builder, and tf-a-coverity
simultaneously. Since tf-a-main is the most high-intensity job and the
high workload would cause the communication channel between Jenkins
master node and the Windows build nodes terminated, occasionally.
move the tf-a-windows-builder to the first job, and run the rest jobs
after it gets done to over come the issue.

Signed-off-by: Arthur She <arthur.she@linaro.org>
Change-Id: I92994b8af629afc80bf182c2230a27e893a5d254