tf-a-builder: Move test submission retries to submit-test-job.sh
We have multiple test facilities to submit tests to, and retry policy
depends on facility. So, keep that logic contained in the submission
script for finer control.

For example, we want to keep existing retry policy for LAVA based on
historical patterns, but want to start with "tabula rasa" with TuxSuite
(no retries, and see how it goes).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I788c1271c47b24ad672243392fd3e733be7de59a