tf-a-lts-release-mail: Support multiple LTS branches and SANDBOX_RUN feature
With --latest flag passed, the job now can work against multiple LTS
branches, by specifying which exactly branch to use.

Also, add SANDBOX_RUN boolean (checkbox) param, to send email to a
different set of recipients. The idea that in a sandbox run (default),
email will go to just maintainers to check that it's content is good,
and then in normal run, it will go to the announcement mailing list
or similar. (As of now, both recipient lists are internal-only).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I912a736a08d5e52989a31cdf0e678b2e09ef0b83