tf-a-commitlint: use `npm ci` instead of `npm install`
The current command (`npm install`) does not seem to be adhering to the
NPM package lockfile for some reason (even though it should, as far as I
can tell). Instead, use `npm ci` to do a clean install of all of the NPM
packages, strictly adhering the lockfile.

Note that while `ci` is intended for use in Continuous Integration
environments, it actually stands for "clean install".

Change-Id: Id9224d2029131daf5e02b202f151aed0d7b35fa3
Signed-off-by: Chris Kay <chris.kay@arm.com>