tf-a-lts-docs-watcher.sh: Add support for sandbox and test projects
With this change, there's support for "sandbox" RTD project to publish docs
to, to serve as a preview before publishing to real production docs for
wide audience.

Additionally, there's a support for fork test project
(sandbox/pfalcon/trusted-firmware-a). The idea behind that project is
to create test releases without affecting production Git repository in
any way. Docs from such a test repo then needs to be published for
review either, hance support in this script.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I199eb2fdc76fa132fc576098ad1a99aa95226c53