tf-a-lts-docs-watcher.sh: Add wait_for_build() function
Currently, the script just ensures that a new release tag is activated
on the RTD side, which schedules a build of the docs. But it finished
otherwise, while the actual build may take a while (10 mins or more).
So, after the job completes, the new docs may be not available for
some time, which may lead to confusion from the users. To be able to
address this concern, add a function to inventory to be able to wait
for a build of a particular doc version to finish. With this change,
this function isn't actually called, that's left for future updates.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I5a9154026a64d2ee4591bfbcd9a69c31531901bc