tf-a-builder/builders.sh: Remove unconditional call to "armclang --vsn"
While we just request to print version and licensing information there,
this command will actually fail if the license is not available. This
is going to be the case now, as we're removing fallback FlexNet license
from jobs. Might have added " || true" to the command, but it will print
a lot of scary warnings if run w/o a license, and it actually was added
to debug UBL activation - we first ran it before activation and then
after. So now, just call for "after" remains. (With activation happening
only or builds actually using ArmClang, while the above command ran
unconditionally, so any issues with it broke absolutely all builds).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I0ee43659193e7ed66d7e96568113a9f6859752ef