tfa-next: add initial ci
Initial job config to run the tfa-next tests. Triggered automatically on
new patchsets uploaded to Gerrit.

For now gateway and builder yamls are direct copies of the non-suffixed
`-tfa-next` versions, just renamed to retain the same Jenkins config
that's applied via the suffix. At some point in the future they can be
replaced with the normal non-suffixed versions.

Change-Id: I2d194aff6760760a72050da21c32b574157ca3c7
Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>