tfa-next: trigger on RUN_CI Gerrit message
Sometimes a manual trigger is required e.g. CI infra was unstable/failed
or failed to trigger on a new patchset.

Manual CI runs via Jenkins do not leave Verified+1 or SUCCESS/FAILURE messages. Patches need to be triggered by the Gerrit trigger plugin for the messages and votes to work.

Add a new regex trigger to start CI when the exact comment "RUN_CI" is
left.

In addition the gerrit-build.*code-review-value lines are not used or
required. The Gerrit plugin should not be setting or touching these
values.

Change-Id: I9a62d983a83a0a1c4e56d6065d2c25ccca3232db
Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>