next: include trivial rebase for run jobs
A new patch set is considered to be trivial rebase if the commit
message is the same as in the previous patch set and if it has
the same diff (including context lines) as the previous patch set.

Edge cases where there is a rebase with no conflicts in the diff
but there are changes involved that change a dependency or a
renaming elsewhere can cause issues in the code nevertheless.

 * set `exclude-trivial-rebase: false` to test on these edge
   cases as well.

Change-Id: Ia406556d89a1f91efbce8aa58bc7fe3276df6654
Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>