ci: archive artefacts even on build failure
The `tf-a-builder` job currently neglects to archive build artefacts if
the job fails, as the artefact directory is not copied to the job
workspace until *after* the script has exited with a failure.

Rather than copying the artefacts directory to the workspace, this
change instead symlinks the artefacts directory to where Jenkins expects
it early on, such that they are always available to Jenkins regardless
of where the build fails.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: I3431c086521bd7870bcc251bb7f33f443897714b