tfa-next: archive artefacts even on build failure
The `tf-a-builder-tfa-next` job currently neglects to archive build
artefacts if the job fails, as the artefact directory is not copied
to the job workspace until *after* the script has exited with a
failure.

 * mimic what the `tf-a-builder` job currently does.

Rather than copying the artefacts directory to the workspace, symlink
 the artefacts directory to where Jenkins expects it early on,
such that they are always available to Jenkins regardless of where
the build fails.

Please see 2d92fdbbc3bc76cbbb6c0ac129fa40bf6c2a2ec9 for more
information.

Change-Id: Id802a7a91f94e01f9e065ca7c7ac13cdac029b2b
Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>