fix: split submodule init and checkout steps
Git does not initialize submodules during `git checkout`, even when
`--recurse-submodules` is used. This causes issues when switching to a
commit that introduces a new submodule, as Git tries to update the
uninitialised submodule [1]. Split the steps to explicitly initialize
submodules after checkout to ensure all nested repositories are properly
loaded.

[1] https://stackoverflow.com/a/72275916

Change-Id: I20a9cf43affd45c0a67ab1ae9a4db0e6f75973a7
Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>