ci: compress repos before storing them on the shared volume
At the moment we're cloning repositories directly into the shared
volume, then copying back out from it repeatedly. This shared volume is
an EFS mount, so that's a pretty slow and expensive process.

Instead, clone each repository to a local directory, tar it up, and then
copy it to the shared volume for use by child jobs. Consuming jobs can
instead untar the file to their working directory.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: Ife6ce8a2069c9dc47899abc742a449d606e53cf5