ci: remove all traces of SQUAD
During the migration to `ci.trustedfirmware.org` from the Arm internal
CI, SQUAD was introduced to do whatever SQUAD does. We struggled to find
a use for it, and eventually reverted to interfacing directly with LAVA
a few years ago.

This change removes all of the now-dead code that used to handle LAVA
job submission via SQUAD.

Change-Id: I8fee32b422bdca1f382ea4bec28c03d0155200ad
Signed-off-by: Chris Kay <chris.kay@arm.com>