ci: configure default Gerrit branch/refspec for static checks
The static checks do not allow you to specify the Gerrit refspec, and
the default branch is empty. This means that the script has to fill
these defaults in instead.

There's really no reason to do that - we can specify reasonable defaults
at the job level, so let's do that.

This also has the benefit of exposing the `GERRIT_REFSPEC` parameter
when manually building the job, which is critical for manual testing.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: I53bf17d31b6f66fed5532e40d252351704df8b96