rfa: configure default Gerrit branch/refspec for static checks
The static checks do not allow you to specify the Gerrit refspec, and
the default branch is empty. This means that the script has to fill
these defaults in instead.

There's really no reason to do that - we can specify reasonable defaults
at the job level, so let's do that.

This also has the benefit of exposing the `GERRIT_REFSPEC` parameter
when manually building the job, which is critical for manual testing.

Change-Id: Ib2ca53d5de60f2d4f84d0e7c0f6d9b47bb836dff
Signed-off-by: Chris Kay <chris.kay@arm.com>
Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>