ci: always clone proper ref
The `clone.sh` currently clones the HEAD of each repository. This is
undesirable for shallow clones, because when the commit under test is
checked out then the HEAD commit is left grafted. It then becomes very
difficult for the job to unshallow the repository, because if this
grafted commit is present in the checked out history then Git fails out
with an obscure error about unreachable references.

Rather than cloning the HEAD ref, instead initialise the repository and
check out the commit under test directly.

Change-Id: I65333b2141ff47d23100701a552b081f30a308df
Signed-off-by: Chris Kay <chris.kay@arm.com>